+++ sync-for-expenses/CodatSyncExpenses/Models/Webhooks/SyncCompleteResponse.cs
@@ -8,7 +8,7 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatSyncExpenses.Models.Webhooks
+namespace Codat.Sync.Expenses.Models.Webhooks
 {
     using System.Net.Http;
     using System;
@@ -29,6 +29,6 @@ namespace CodatSyncExpenses.Models.Webhooks
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
     }
 }